dashboardbackend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.1
psycopg2-binary==2.9.9
pymysql==1.1.0
cryptography==41.0.7
apscheduler==3.10.4